service base

+++ manager/src/main/java/it/andreamugellini/manager/srvs/SrvManager.java
@@ -6,40 +6,66 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
+import it.andreamugellini.manager.ctrls.dto.DTOCalculation;
+import it.andreamugellini.manager.exc.ExCalculationNotFound;
 import it.andreamugellini.manager.rps.RepoCalculation;
 import it.andreamugellini.manager.rps.entity.EntCalculation;
 
 @Service
 public class SrvManager {
-	
-	@Autowired
-	private  KafkaTemplate<String, EntCalculation> kafka;
-	
-	@Autowired
-	private  RepoCalculation repoCalculation;
-	
-	
 
-	public EntCalculation create(EntCalculation calculation) {
-		
-		EntCalculation c = this.repoCalculation.save(calculation);
-		kafka.send("evaluate-calc", calculation);
-		return c;
+	@Autowired
+	private KafkaTemplate<String, EntCalculation> kafka;
+
+	@Autowired
+	private RepoCalculation repoCalculation;
+
+	public DTOCalculation create(DTOCalculation dto) {
+
+		EntCalculation c = new EntCalculation();
+		c.setOperation(dto.getOperation());
+		this.repoCalculation.save(c);
+		kafka.send("evaluate-calc", c);
+		dto.setId(c.getId());
+
+		return dto;
 	}
 
-	public EntCalculation get(Long id) {
+	public DTOCalculation get(String id) {
 		Optional<EntCalculation> c = this.repoCalculation.findById(id);
-		return c.get();		
+		if (!c.isPresent())
+			throw new ExCalculationNotFound(id);
+
+		DTOCalculation dto = new DTOCalculation();
+		dto.setId(c.get().getId());
+		dto.setOperation(c.get().getOperation());
+		dto.setResult(c.get().getResult());
+
+		return dto;
+
 	}
 
-	public EntCalculation update(Long id, EntCalculation calculation) {
-		this.repoCalculation.save(calculation);
-		kafka.send("calculate", calculation);
-		return calculation;
+	public DTOCalculation update(String id, DTOCalculation dto) {
+
+		Optional<EntCalculation> c = this.repoCalculation.findById(id);
+		if (!c.isPresent())
+			throw new ExCalculationNotFound(id);
+
+		c.get().setOperation(dto.getOperation());
+
+		this.repoCalculation.save(c.get());
+
+		kafka.send("calculate", c.get());
+
+		dto.setId(id);
+		return dto;
 	}
 
-	public void delete(Long id) {
+	public void delete(String id) {
+		Optional<EntCalculation> c = this.repoCalculation.findById(id);
+		if(!c.isPresent()) throw new ExCalculationNotFound(id);
 		this.repoCalculation.deleteById(id);
+
 	}
 
 }
